models: refactor model folder inside mlpipe

+++ mlpipe/utils.py
@@ -2,4 +2,4 @@ import numpy as np
 
 def to_categorical(y, num_classes):
     """ 1-hot encodes a tensor """
-    return np.eye(num_classes, dtype='uint8')[y]
+    return np.eye(num_classes, dtype='int')[y.astype('int')]
